+++ Runtime/BidscubeSDK/Controllers/ConsentTestScene.cs
@@ -9,7 +9,7 @@ namespace BidscubeSDK.Controllers
     /// Consent Test Scene - Consent form and consent management testing
     /// Based on iOS ConsentTestView
     /// </summary>
-    public class ConsentTestScene : MonoBehaviour, IAdCallback, IConsentCallback
+    public class ConsentTestScene : MonoBehaviour, IAdCallback, IRewardedAdCallback, IConsentCallback
     {
         [Header("SDK Configuration")]
         [SerializeField] private string _placementId = "20212";
@@ -374,6 +374,11 @@ namespace BidscubeSDK.Controllers
         public void OnVideoAdSkippable(string placementId)
         {
             LogMessage($"Video ad skippable: {placementId}");
+        }
+
+        public void OnUserRewarded(string placementId)
+        {
+            LogMessage($"User rewarded: {placementId}");
         }
 
         public void OnInstallButtonClicked(string placementId, string buttonText)